Refactor tests into groups
I bullied GPT into writing me a crappy script to switch the old
`findfirst` tests to use `match`, which is the only interface which
actually matters. Found a bug in the process where I was making
SubString views without properly checking for the previous index,
reminding myself that if code isn't tested on multibyte Unicode
it's assuredly broken with respect to it.