Tests... pass
I demoted an error to a warning because it isn't a problem for
the rest of that test and is in an aspect of the compiler that
needs a revamp anyway.

That said, keying by instruction is out of the picture, so I can
get that working the way I want it to.