Move the capture tag to the PCapture itself
This is starting to be right; I'm still losing a key somewhere
but I'm thinking it doesn't matter.

Now that I've hit on tagging each capture instruction there's no
need to use a "capvec" or collect it, because a) I'm not getting
dict clobbers and b) I don't have to recreate the exact
instruction for retrieval, so I can do the other optimizations,
just have to keep the tag right and it'll work out.