Get rid of all that awful mergeaux! nonsense
Much cleaner to walk the Pattern tree and _look_ for what we need,
when we need it, rather than to bubble it up and stick it into
every intermediate level, whether it can use it or not.

When I'm rewriting grammar compiling I can go through and make the
aux field a Union{Dict,Nothing} and only create the ones which get
`prepare!`'ed.