Something putatively accomplished
This... wasn't a total waste of a day's effort. I now walk the
grammar recursively, which I'm currently only using for layout
of code, which actually doesn't matter at all, but which can be
used for left-recursion and other property detection eventually.

I've been rushing through with a New Years hangover, it's time to
do other things.