Much mulling over headfails
The rule for TSeq in lplcode.c makes no sense to me, I've thought
about this for hours. But the other tests are on solid ground and
I want to add the rest of the "by the book" optimizations, so I'm
going to just take it for granted. The worst that can happen is
that it reports a not-headfailing pattern as headfail, after all.