Add virtual offset property; Cposition is Int
I discovered a most unwelcome issue with SubStrings (filed a bug
with Julia), which is that empty SubStrings do not contain the
offset they were set to, but rather, 0.

I may put together a package based on the latest Julia SubString
code, which corrects this for my own purposes, and as a proof of
concept to Julia that it can work.  "Fun" little project.