Refactor PegMatch with Bool .full
This makes more sense for us, especially once I add a SubString
variant which retains the position of an empty string.

PEGs always match from the first character, and are more commonly
used to match an entire string. A SubString can be interrogated as
to its position within the String, if the user has need of that
last matched character, for whatever reason.