Better behavior inside predicates
We don't want to remove marks inside predicates, or a check can't
be used for both lookahead and matching, which is rather common.

Also fixes buggy behavior for zero-length matches, which, Julia
indexing being what it is, put the end point one character before
the start point, meaning the necessary -1 has to be in the substring,
not the offset, if we're to use a `UInt` to store it, as we should.