No more global state with checks
The circumstances under which this could ever matter are somewhat
farfetched but this is The Right Way To Do It, and I don't like
global state as a matter of principle.  This makes the dicts which
map tags to symbols or functions into a cache, and I don't think
I actually need them any longer.