Add "firstset" function
The code for TestSets is devilishly hard to follow, because in C
you end up passing in and mutating a bunch of references, and the
tradition is then to use the return value to signal various sorts
of state.

I'm not convinced that "firstset, followset" is the right way for
us to implement this optimization.