Move peephole! to prepare!, call on all Patterns
This also changes the Grammar code so that it, properly, jumps to
the end, instead of hard-coding OpEnd as the second instruction.
This is important, because it allows Grammars to be embedded in
other Grammars, something we will eventually be doing.

Then the peephole goes and optimizes the Jump into an End.  I
need to give more thought to how pattern combination works, because
the existing code is (now at least) broken. A Choice, for instance,
if peepholed, the Jumps to End will be optimized into End, but
this is a problem if that Choice is later combined into a bigger
pattern.

I think the answer is to stop auto-compiling the auxiliaries,
with the whole spurious `hoist!`, `compile!`, `_compile!`
distinction. There should be one command which is "give me your
bytecode", and another one which is "prepare this pattern for
matching", and they aren't related at all: a prepared pattern
will freshly compile the bytecode and hand it over, in unoptimized
form, when 'compiled'.

This has the right properties. An enclosed pattern can mutate the
Vector it's given, possibly just reusing it, because it knows
there's no reference being kept, and a prepared pattern can be
embedded in another pattern correctly; it doesn't lose its
prepared form, but it also doesn't reference it when the enclosing
pattern compiles.  This also means that sub-patterns will have
empty code pointers instead of full copies that they're never
going to use. That's a pretty small thing, the extra allocation,
but it's the principle of it.