1000x less allocation with parameterized VMState
Because VMState is mutable, the compiler must assume that the
concrete type of .subject can change at any time.  By using a
parameter, it is now an error to substitute, say, a SubString
for a VMState set up for a String. As a result, the code doesn't
need to box characters or Int pointers, and allocation is down to
a reasonable amount.