Replace _compile! with build
The Julia convention is that a `!` means the method will modify
the first argument.  In this case the mutation happens to an
optional parameter which is constructed ad hoc if not provided,
so no bang.

The codebase is currently unusable and all these modifications are
untested.  The next commit will finish the job.