Begin adding set vectors as pseudoinstruction
When this is complete, every item in an instruction Vector will be
the same size.  I don't believe Julia will use this information to
inline instructions, because it can't: someone could define a
new subclass of Instruction of a different width and try to add it
to a Vector, so they have to be boxed.  Aside from that, they would
at minimum need to be boxed with type information, and that isn't
what we want either.

But it does mean we can use some `reinterpret` trickery to pack
them into a Vector described in terms of one of the primitive types
(this is what UrInst was added to do) and that *should* unbox them.

I saw some notes in the discourse about how to verify that a Vector
is inlining isbits structs, which I'll need to find before we get
to the truly fun part, making the dispatch loop type stable.