Fix U8 bug
Truly embarrassed that I didn't add a test of failure, which
would have caught this immediately.  ByteInst (now named
LeadByteInst, as it should be) is a jumping instruction, not a
pass/fail.  With the conversion of CharInst to use bytes, U8 is
simple to implement as a character, the validity doesn't make a
difference to the VM.