Refactors PSet | PChar and vice versa
This was an unsuccessful attempt to work around a bug in the
UnitRangesSortedSets library, and in fact, the original code
reduces to this in practice.  This version is perhaps more
efficient than the previous, in a way which is unlikely to ever
matter to anyone, but there being no compelling reason to revert
it, the change stays.