Move capture sorting and matching to new file
I prefer to have the VM be its own thing, with the various pre-
and post- actions living in their own file.

One, the VM will be changing quite a bit soon, two, the surface
area of matching is expected to increase quite a bit.  It's a
logical separation of concerns.