Better errors for invalid rule names
I've forgotten that rulenames have to be a quoted symbol a couple
times.  It would be possible to allow symbols, after all, there is
a deliberate ambiguity between the rulename and the variable name,
because :rule_name becomes the variable rule_name.

But it's better to throw an error rather than follow Postel's Law
here, because this means that rules can be copied over to grammars
without any changes, along with the other advantages of a
consistent syntax, like, um.  Consistency.