Add an instrumented match/vm loop
Probably not the only one of its type, just the easiest.

Instrumenting backtracks is probably more helpful than just
counting instructions, and it would be fun/useful to display a
heat map of the string, showing where the pattern backtracks the
most.

That's fairly well involved, but VM loops are cheap until I finish
the type stability project, and we can keep the old ones with a
bit of squeezing: we can retrieve or reconstruct the original code,
and write a dispatcher that works like the one which will be in the
main dispatch loop by then.